+++ src/libs6rc/s6rc_servicedir_block.c
@@ -1,5 +1,6 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <errno.h>
 #include <unistd.h>
 #include <skalibs/bytestr.h>
@@ -9,7 +10,7 @@
 
 int s6rc_servicedir_block (char const *dir)
 {
-  unsigned int dirlen = str_len(dir) ;
+  size_t dirlen = str_len(dir) ;
   s6_svstatus_t status ;
   char fn[dirlen + 6] ;
   if (!s6_svstatus_read(dir, &status)) return -1 ;
