+++ src/libs6rc/s6rc_servicedir_unblock.c
@@ -1,5 +1,6 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <errno.h>
 #include <unistd.h>
 #include <skalibs/bytestr.h>
@@ -10,7 +11,7 @@ int s6rc_servicedir_unblock (char const *dir, int h)
 {
   if (h)
   {
-    unsigned int dirlen = str_len(dir) ;
+    size_t dirlen = str_len(dir) ;
     char fn[dirlen + 6] ;
     byte_copy(fn, dirlen, dir) ;
     byte_copy(fn + dirlen, 6, "/down") ;
